Race class moved into the leader_board script, consistent with style
